kafka_pipeline/KafkaStreamPoC/requirements.txt
kafka-python
pyspark
faiss-cpu 
sentence-transformers
langchain
langchain-experimental
textract==1.6.3
pdfplumber
tqdm
numpy
streamlit
pandas
openpyxl
pytesseract
pillow
torchvision
transformers
docker

